refactor(Address): :recycle: Add modal for flow details
