+++ src/core/components/ViewDetailsCell/ViewDetailCell.interfaces.ts
@@ -1,3 +1,5 @@
 export interface ViewDetailCellProps {
-  link: string;
+  link?: string;
+  value?: string;
+  onClick?: Function;
 }
