+++ src/core/components/ViewDetailsCell/index.tsx
@@ -1,12 +1,22 @@
+import { useCallback } from 'react';
+
 import { SearchIcon } from '@patternfly/react-icons';
 import { Link } from 'react-router-dom';
 
 import { ViewDetailCellProps } from './ViewDetailCell.interfaces';
 
-const viewDetailCell = ({ link }: ViewDetailCellProps) => (
-  <Link to={link}>
-    <SearchIcon />
-  </Link>
-);
+const ViewDetailCell = function ({ link, value, onClick }: ViewDetailCellProps) {
+  const handleOnClick = useCallback(() => {
+    if (onClick) {
+      onClick(value);
+    }
+  }, [value, onClick]);
 
-export default viewDetailCell;
+  return (
+    <Link to={link || ''} onClick={handleOnClick}>
+      <SearchIcon />
+    </Link>
+  );
+};
+
+export default ViewDetailCell;
