+++ src/pages/Addresses/routes.tsx
@@ -4,7 +4,6 @@ import { AddressesRoutesPaths } from './Addresses.enum';
 
 const Addresses = lazy(() => import(/* webpackChunkName: "addresses" */ './views/Addresses'));
 const FlowPairs = lazy(() => import(/* webpackChunkName: "addresses-flow-pairs" */ './views/FlowPairs'));
-const FlowsPair = lazy(() => import(/* webpackChunkName: "addresses-flow-pair" */ './views/FlowPair'));
 
 export const addressesRoutes = [
   {
@@ -14,9 +13,5 @@ export const addressesRoutes = [
   {
     path: `${AddressesRoutesPaths.Addresses}/:address`,
     element: <FlowPairs />
-  },
-  {
-    path: `${AddressesRoutesPaths.Addresses}/:address/:flowPairId`,
-    element: <FlowsPair />
   }
 ];
