+++ src/pages/shared/FlowPairs/FlowPairs.enum.ts
@@ -25,3 +25,22 @@ export enum FlowPairsColumnsNames {
   StatusCode = 'Status',
   Trace = 'Trace'
 }
+
+export enum FlowLabels {
+  Protocol = 'Protocol',
+  Process = 'Process',
+  Host = 'Source Host',
+  Port = 'Source Port',
+  Method = 'Method',
+  Trace = 'Trace',
+  DestHost = 'Destination Host',
+  DestPort = 'Destination Port',
+  Flow = 'Client',
+  CounterFlow = 'Server',
+  ByteRate = 'Byte rate',
+  BytesTransferred = 'Bytes transferred',
+  ByteUnacked = 'Byte unacked',
+  WindowSize = 'Window Size',
+  Latency = 'Latency',
+  Duration = 'Duration'
+}
